NXP backend: Rename post-processing pass directory.
