+++ backends/nxp/backend/ir/neutron_ir_post_processing/optimizations/prune_transpose_operators.py
@@ -1,4 +1,4 @@
-# Copyright 2024-2025 NXP
+# Copyright 2024-2026 NXP
 #
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
@@ -9,21 +9,21 @@ from executorch.backends.nxp.backend.ir.converter.conversion.translator import (
     apply_permutation_to,
     combine_permutations,
 )
-from executorch.backends.nxp.backend.ir.tflite_generator import tflite_model
-from executorch.backends.nxp.backend.ir.tflite_optimizer.optimizations.base_optimization import (
+from executorch.backends.nxp.backend.ir.neutron_ir_post_processing.optimizations.base_optimization import (
     BaseOptimization,
 )
-from executorch.backends.nxp.backend.ir.tflite_optimizer.pattern_matcher import (
+from executorch.backends.nxp.backend.ir.neutron_ir_post_processing.pattern_matcher import (
     MultipleSameOps,
     Op,
     PatternMatcher,
 )
-from executorch.backends.nxp.backend.ir.tflite_optimizer.tensor_rules import (
+from executorch.backends.nxp.backend.ir.neutron_ir_post_processing.tensor_rules import (
     RuleOr,
     TensorHasData,
     TensorIsNotModelOutput,
     TensorsHaveData,
 )
+from executorch.backends.nxp.backend.ir.tflite_generator import tflite_model
 from executorch.backends.nxp.backend.neutron_operator_support import (
     transposition_is_supported_on_neutron,
 )
